pyproject.toml
[project]
name = "audiotoaudio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-genai>=1.29.0",
    "google-generativeai>=0.8.5",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.1.1",
    "pyttsx3>=2.99",
    "speechrecognition>=3.14.3",
]
